+++ Editor/Drawers/HandleDrawers/DrawWireDiscDrawer/DrawWireDiscAttributeDrawer.cs
@@ -57,7 +57,7 @@ namespace SaintsField.Editor.Drawers.HandleDrawers.DrawWireDiscDrawer
             Type rawType = wireDiscInfo.MemberInfo is FieldInfo fi
                 ? fi.FieldType
                 : ((PropertyInfo)wireDiscInfo.MemberInfo).PropertyType;
-            Type fieldType = wireDiscInfo.SerializedProperty.isArray? ReflectUtils.GetElementType(rawType): rawType;
+            Type fieldType = SerializedUtils.IsArrayOrDirectlyInsideArray(wireDiscInfo.SerializedProperty)? ReflectUtils.GetElementType(rawType): rawType;
             bool filedIsNumber = fieldType == typeof(float) || fieldType == typeof(double) || fieldType == typeof(int) || fieldType == typeof(long) || fieldType == typeof(short);
 
             if(filedIsNumber)