+++ Samples~/Scripts/SaintsEditor/Testing/ShowInInspectorEditing.cs
@@ -26,33 +26,33 @@ namespace SaintsField.Samples.Scripts.SaintsEditor.Testing
             private MyEnum _myEnum;
         }
 
-        [ShowInInspector] private bool _boolV;
-        [ShowInInspector] private byte _byteV;
-        [ShowInInspector] private sbyte _sbyteV;
-        [ShowInInspector] private short _shortV = -5;
-        [ShowInInspector] private ushort _ushortV = 0;
-        [ShowInInspector] private long _longV = long.MaxValue;
-        [ShowInInspector] private ulong _uLongV = ulong.MaxValue;
-        [ShowInInspector] private float _floatV = 0.1234f;
-        [ShowInInspector] private double _doubleV = 0.1234d;
-        [ShowInInspector] private string _stringV = "Hi";
-        [ShowInInspector] private Vector2 _vector2V = Vector2.one;
-        [ShowInInspector] private Vector3 _vector3V = Vector3.one;
-        [ShowInInspector] private Vector4 _vector4V = Vector4.one;
-        [ShowInInspector] private Vector2Int _vector2VInt = Vector2Int.one;
-        [ShowInInspector] private Vector3Int _vector3VInt = Vector3Int.one;
-        [ShowInInspector] private Color _colorV = Color.red;
-        [ShowInInspector] private Bounds _boundsV = new Bounds(Vector3.one, Vector3.up);
-        [ShowInInspector] private Rect _rectV = new Rect(0, 0, 1, 1);
-        [ShowInInspector] private RectInt _rectIntV = new RectInt(0, 0, 1, 1);
-        [ShowInInspector] private MyEnum _myEnum;
-        [ShowInInspector] private GameObject _go;
-        [ShowInInspector] private Transform _trans;
-        [ShowInInspector] private Scriptable _so;
+        [ShowInInspector, Ordered] private bool _boolV;
+        [ShowInInspector, Ordered] private byte _byteV;
+        [ShowInInspector, Ordered] private sbyte _sbyteV;
+        [ShowInInspector, Ordered] private short _shortV = -5;
+        [ShowInInspector, Ordered] private ushort _ushortV = 0;
+        [ShowInInspector, Ordered] private long _longV = long.MaxValue;
+        [ShowInInspector, Ordered] private ulong _uLongV = ulong.MaxValue;
+        [ShowInInspector, Ordered] private float _floatV = 0.1234f;
+        [ShowInInspector, Ordered] private double _doubleV = 0.1234d;
+        [ShowInInspector, Ordered] private string _stringV = "Hi";
+        [ShowInInspector, Ordered] private Vector2 _vector2V = Vector2.one;
+        [ShowInInspector, Ordered] private Vector3 _vector3V = Vector3.one;
+        [ShowInInspector, Ordered] private Vector4 _vector4V = Vector4.one;
+        [ShowInInspector, Ordered] private Vector2Int _vector2VInt = Vector2Int.one;
+        [ShowInInspector, Ordered] private Vector3Int _vector3VInt = Vector3Int.one;
+        [ShowInInspector, Ordered] private Color _colorV = Color.red;
+        [ShowInInspector, Ordered] private Bounds _boundsV = new Bounds(Vector3.one, Vector3.up);
+        [ShowInInspector, Ordered] private Rect _rectV = new Rect(0, 0, 1, 1);
+        [ShowInInspector, Ordered] private RectInt _rectIntV = new RectInt(0, 0, 1, 1);
+        [ShowInInspector, Ordered] private MyEnum _myEnum;
+        [ShowInInspector, Ordered] private GameObject _go;
+        [ShowInInspector, Ordered] private Transform _trans;
+        [ShowInInspector, Ordered] private Scriptable _so;
         // private void SetDummy() => _dummy = _so;
 
-        [ShowInInspector] private MyClass _myClass;
-        [ShowInInspector] private MyClass _myClassD = new MyClass
+        [ShowInInspector, Ordered] private MyClass _myClass;
+        [ShowInInspector, Ordered] private MyClass _myClassD = new MyClass
         {
             MyString = "Hi",
         };
@@ -62,9 +62,9 @@ namespace SaintsField.Samples.Scripts.SaintsEditor.Testing
             public string MyString;
         }
 
-        [ShowInInspector] private MyStruct _myStruct;
+        [ShowInInspector, Ordered] private MyStruct _myStruct;
 
-        [ShowInInspector] private Color[] _colors = {Color.red, Color.green, Color.blue};
+        [ShowInInspector, Ordered] private Color[] _colors = {Color.red, Color.green, Color.blue};
         [ShowInInspector, Ordered] private Color[] _colorEmptyArray;
 
         [Button, Ordered]
@@ -101,7 +101,7 @@ namespace SaintsField.Samples.Scripts.SaintsEditor.Testing
 
         [ShowInInspector, Ordered] private MyClass[] _myClasses;
 
-        [ShowInInspector] private Dictionary<string, Color> _dictColors = new Dictionary<string, Color>
+        [ShowInInspector, Ordered] private Dictionary<string, Color> _dictColors = new Dictionary<string, Color>
         {
             { "Red", Color.red },
             { "Green", Color.green },
@@ -125,7 +125,7 @@ namespace SaintsField.Samples.Scripts.SaintsEditor.Testing
             public int MyInt { get; set; }
         }
 
-        [ShowInInspector] private static Inter _inter;
+        [ShowInInspector, Ordered] private static Inter _inter;
 
         public class GeneralDummyClass: IDummy
         {
@@ -139,9 +139,9 @@ namespace SaintsField.Samples.Scripts.SaintsEditor.Testing
             public string GenDumString;
         }
 
-        [ShowInInspector] private static IDummy _dummy;
+        [ShowInInspector, Ordered] private static IDummy _dummy;
 
-        [Button]
+        [Button, Ordered]
         private void DebugDummy() => Debug.Log(_dummy);
 
 #if UNITY_EDITOR
@@ -163,11 +163,7 @@ namespace SaintsField.Samples.Scripts.SaintsEditor.Testing
 
         [NonSerialized, ShowInInspector] public NullSwap _nullSwap;
 
-        [ShowInInspector]
-        private IReadOnlyDictionary<int, string> _readOnlyDict =
-            new Dictionary<int, string>();
-
-        [ShowInInspector]
+        [ShowInInspector, Ordered]
         private IEnumerable<int> _ie = Enumerable.Range(0, 3);
 
         private class MyList : IReadOnlyList<int>
@@ -194,12 +190,30 @@ namespace SaintsField.Samples.Scripts.SaintsEditor.Testing
             public int this[int index] => Lis[index];
         }
 
-        [ShowInInspector] private MyList _myListNull;
-        [ShowInInspector] private MyList _myListSome = new MyList { Lis = new[] { 1, 2 } };
+        [ShowInInspector, Ordered] private MyList _myListNull;
+        [ShowInInspector, Ordered] private MyList _myListSome = new MyList { Lis = new[] { 1, 2 } };
 
-        [ShowInInspector] public Dictionary<MyEnum, int> MyDictionary = new Dictionary<MyEnum, int>
+        [Button, Ordered]
+        private void DictExternalAdd()
+        {
+            _myDictionaryNull["External"] = 1;
+        }
+        [ShowInInspector, Ordered] private Dictionary<string, int> _myDictionaryNull;
+
+
+        [ShowInInspector, Ordered] public Dictionary<MyEnum, int> MyDictionary = new Dictionary<MyEnum, int>
         {
             {MyEnum.One, 1},
         };
+
+        [ShowInInspector, Ordered]
+        private IReadOnlyDictionary<int, string> _readOnlyDict =
+            new Dictionary<int, string>
+            {
+                {1, "One"},
+                {2, "Two"},
+            };
+
+        [ShowInInspector, Ordered] public Dictionary<int, int[]> _intToInts = new Dictionary<int, int[]>();
     }
 }